+++ src/TurtlePath.FluentValidation/FluentValidationTurtlePathBuilderExtensions.cs
@@ -19,7 +19,7 @@ namespace TurtlePath.FluentValidation
             if (builder == null)
                 throw new ArgumentNullException(nameof(builder));
 
-            builder.Services.TryAddSingleton<IValidatorAdapter, ValidatorAdapter>();
+            builder.Services.TryAddSingleton<IValidatorAdapter, FluentValidationAdapter>();
 
             return builder;
         }